upgraded typescript extended dual modujle type support updated tailwind to v4 and improved support: * No more endless watch cycles being triggered. * Lazy imported bundles are also included in tailwinds css generation * We turn off tailwind preflight, but then recreate it almost entirely. But without destructive/strict selectors that ended up overwriting styles from CSS modules * imported as first postcss plugin so that styles & theme variables from tailwind come first in the bundle.
SCSS - removed entirely. we rely on postcss for things like nesting.

CSS modules
- are now only applied to css files with .module.css names
- imrpoved css loader for css modules inside node.js

improved option to see which files changed and triggered a rebuild (to be cleaned up a bit)

config-webpack will likely be outdated, if its used to bundle a lincd package (we dont really do that at the moment)